quanticx/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
aiohttp==3.10.10
pydantic==2.9.1
python-dotenv==1.0.1
SQLAlchemy==2.0.35
aiosqlite==0.20.0
redis==5.0.7
slowapi==0.1.9
monobank
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
celery==5.4.0
prometheus-client==0.20.0
sentry-sdk==2.13.0
python-json-logger==2.0.7
alembic==1.13.2
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-exporter-otlp==1.26.0
reportlab==4.2.2
